==> external-validation/panel-sequencing/0.prep-other-data-sources.ipynb <==
import pandas as pd
import numpy as np

---

sample = pd.read_csv('msk_chord_2024/data_clinical_sample.txt', sep='\t', comment='#')
sample['GENE_PANEL'].value_counts()

----

gene_list = pd.read_csv('misc-data/cancerGeneList.tsv', sep = '\t')
gene_list[gene_list['MSK-IMPACT'] == 'Yes'].shape[0]

gene_list[gene_list['MSK-HEME'] == 'Yes'].shape[0]

this is a 505 gene list, the latest panel. need to find ones that match all panels.
