==> external-validation/panel-sequencing/1-1.curate-validation-cohort.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm
import os
import warnings
warnings.filterwarnings('ignore')

from misc.extract_clinical_data import extract_clinical_data_minimum
from misc.extract_molecular_data import extract_dna_variants, extract_cna, extract_structural_variants
patient_data_path = 'msk_chord_2024/data_clinical_patient.txt'
sample_data_path = 'msk_chord_2024/data_clinical_sample.txt'
snv_data_path = 'msk_chord_2024/data_mutations.txt'
cna_data_path = 'msk_chord_2024/data_cna.txt'
sv_data_path = 'msk_chord_2024/data_sv.txt'

----

patient = pd.read_csv('msk_chord_2024/data_clinical_patient.txt', sep='\t', comment='#')
treatment = pd.read_csv('msk_chord_2024/data_timeline_treatment.txt', sep='\t', comment='#')
sample = pd.read_csv('msk_chord_2024/data_clinical_sample.txt', sep='\t', comment='#')

targeted_tx_patients = treatment[
    #select patients that have received targeted therapies after msk-impact
    (treatment.SUBTYPE == 'Targeted')&(treatment.START_DATE > 0)&
    #select patients that have only had one msk-impact test (for the sake of "gold standard", we'll think of ways to incorporate later)
    (treatment['PATIENT_ID'].isin(sample['PATIENT_ID'].value_counts()[sample['PATIENT_ID'].value_counts() == 1].index))]
print(f'Number of unique patients receiving targeted therapy post test: {len(targeted_tx_patients.PATIENT_ID.unique())}')

#for now, it seems like we don't need to think about multiple samples per patient
sample[sample.PATIENT_ID.isin(targeted_tx_patients.PATIENT_ID.unique())].PATIENT_ID.value_counts()

treatment_filtered = treatment[(treatment.PATIENT_ID.isin(targeted_tx_patients.PATIENT_ID.unique()))&
                               (treatment.START_DATE > 0)]
treatment_filtered.head()

'''
we want to 
1. identify the "first line treatment(s)" received after msk-impact was administered.
2. if the "first line treatment(s)" weren't targeted, filter out for now.
3. if the "first line treatment(s)" were targeted, keep and save the agent(s) administered first. 
'''
first_line_table = []
for patient in tqdm(treatment_filtered.PATIENT_ID.unique()):
    patient_treatments = treatment_filtered[treatment_filtered.PATIENT_ID == patient]
    first_line_treatment = patient_treatments[patient_treatments.START_DATE == patient_treatments.START_DATE.min()]
    agent_type = first_line_treatment.SUBTYPE.unique()
    first_line_timing = first_line_treatment.START_DATE.tolist()[0]
    first_line_treatment_agent = set(first_line_treatment.AGENT.tolist())
    first_line_treatment_type = set(first_line_treatment.SUBTYPE.tolist())
    first_line_table.append([patient, first_line_timing, first_line_treatment_agent, first_line_treatment_type])
first_line_table = pd.DataFrame(first_line_table, columns=['PATIENT_ID', 'FIRST_LINE_TIMING', 'FIRST_LINE_TREATMENT', 'FIRST_LINE_TREATMENT_TYPE'])
first_line_table['TARGETED_THERAPY_GIVEN'] = first_line_table['FIRST_LINE_TREATMENT_TYPE'].apply(lambda x: 'Targeted' in x)
first_line_table['INVESTIGATIONAL_THERAPY_GIVEN'] = first_line_table['FIRST_LINE_TREATMENT_TYPE'].apply(lambda x: 'Investigational' in x)
os.makedirs('reports/filtered-data', exist_ok=True)
first_line_table.to_csv('reports/filtered-data/first_line_treatments_post_msk.csv', index=False)
print(first_line_table.shape)
first_line_table.head()

first_line_table.FIRST_LINE_TIMING.describe()
#realistically, it's probably reasonable to cut off after 1 year

first_line_table = first_line_table[first_line_table.FIRST_LINE_TIMING <= 365]
first_line_table.to_csv('reports/filtered-data/first_line_treatments_post_msk.csv', index=False)
print(first_line_table.shape)
first_line_table.head(1)

sample[sample.PATIENT_ID.isin(first_line_table.PATIENT_ID.unique())].CANCER_TYPE.value_counts()

pd.crosstab(first_line_table.TARGETED_THERAPY_GIVEN, first_line_table.INVESTIGATIONAL_THERAPY_GIVEN)

we might want to remove the patients that were given investigational therapies, because of a lack of information.

first_line_table[first_line_table.INVESTIGATIONAL_THERAPY_GIVEN == False]['FIRST_LINE_TREATMENT_TYPE'].value_counts()

sample[sample.PATIENT_ID.isin(first_line_table[first_line_table.INVESTIGATIONAL_THERAPY_GIVEN == False].PATIENT_ID.unique())].CANCER_TYPE.value_counts()
