==> external-validation/panel-sequencing/1-3.test-formatting.ipynb <==
import pandas as pd
import numpy as np
import os
import warnings
warnings.filterwarnings('ignore')

from misc.extract_clinical_data import extract_clinical_data_minimum
from misc.extract_molecular_data import extract_dna_variants, extract_cna, extract_structural_variants
patient_data_path = 'msk_chord_2024/data_clinical_patient.txt'
sample_data_path = 'msk_chord_2024/data_clinical_sample.txt'
snv_data_path = 'msk_chord_2024/data_mutations.txt'
cna_data_path = 'msk_chord_2024/data_cna.txt'
sv_data_path = 'msk_chord_2024/data_sv.txt'

----
#### rough overview of report content
- basic patient and sample info
    - patient id, age, gender, birth date, 
    - test performed, test description, accession number, sample collection date, path diagnosis, % of tumor cells in sample
- sequencing quality
    - number of total reads, mean reads per taargeted exon, X% of exons with >30 reads
- dna variants
    - tiered list (tier 1-4) of 
        - `gene cHGV (pHGV) exon % of reads, total reads`
        - eg. TP53 c.613T_>C (p.Y205H), exon 2 - in 50% of 73 reads**
    - negative for mutations in genes with clinical relevance for tumor type
- copy number variation
    - list of `location gene copy number type`
    - eg. 1p12 NOTCH2   Low copy number gain
- chromosomal rearrangement

----

patient = pd.read_csv('msk_chord_2024/data_clinical_patient.txt', sep='\t', comment='#')
treatment = pd.read_csv('msk_chord_2024/data_timeline_treatment.txt', sep='\t', comment='#')
sample = pd.read_csv('msk_chord_2024/data_clinical_sample.txt', sep='\t', comment='#')

targeted_tx_patients = pd.read_csv('reports/filtered-data/first_line_treatments_post_msk.csv')
targeted_tx_patients.head(1)

def generate_report(patient_id):
    '''
    Generate a report for a given patient ID by extracting clinical and molecular data.
    '''

    report = ''

    report_text = extract_clinical_data_minimum(
        patient_data_path=patient_data_path,
        sample_data_path=sample_data_path,
        patient_id=patient_id
    )
    report += report_text

    report += '\n============='

    report += '\nDNA Variants:'
    report_text = extract_dna_variants(
        snv_data_path=snv_data_path,
        sample_data_path=sample_data_path,
        patient_id=patient_id
    )
    report += '\n' + report_text

    report += '\nCNA Data:'
    report_text = extract_cna(
        cna_data_path=cna_data_path,
        sample_data_path=sample_data_path,
        patient_id=patient_id
    )
    report += '\n' + report_text

    report += '\nStructural Variants:'
    report_text = extract_structural_variants(
        sv_data_path=sv_data_path,
        sample_data_path=sample_data_path,
        patient_id=patient_id
    )
    report += '\n' + report_text

    return report


#patients = sample[sample.PATIENT_ID.isin(targeted_tx_patients.PATIENT_ID.unique())].PATIENT_ID.tolist()[:20]
'''
os.makedirs('reports/test', exist_ok=True)
with open('reports/test/selected-samples.json') as f:
    import json
    patients = json.load(f)
from tqdm import tqdm
for patient_id in tqdm(patients):
    patient_id = patient_id.split('-T')[0]
    report = generate_report(patient_id)
    with open(f'reports/test/{patient_id}.txt', 'w') as f:
        f.write(report)
'''

os.makedirs('reports/test', exist_ok=True)
'''
with open('reports/test/selected-samples-no-previous-treatment.json') as f:
    import json
    patients = json.load(f)
from tqdm import tqdm
for patient_id in tqdm(patients):
    patient_id = patient_id.split('-T')[0]
    report = generate_report(patient_id)
    with open(f'reports/test/{patient_id}.txt', 'w') as f:
        f.write(report)
'''

cohort = pd.read_csv('reports/test/selected-samples-no-previous-treatment-metadata.csv')
import ast
cohort['FIRST_LINE_TREATMENT_TYPE'] = cohort['FIRST_LINE_TREATMENT_TYPE'].apply(lambda x: ast.literal_eval(x) if pd.notnull(x) else [])
cohort = cohort[cohort.FIRST_LINE_TREATMENT_TYPE.apply(lambda x: 'Targeted' in x)]
patients = cohort.PATIENT_ID.tolist()
from tqdm import tqdm
for patient_id in tqdm(patients):
    report = generate_report(patient_id)
    with open(f'reports/test/{patient_id}.txt', 'w') as f:
        f.write(report)
