==> external-validation/panel-sequencing/9.evaluation-of-cohort.ipynb <==
import pandas as pd
import numpy as np
import os
import ast
import warnings
warnings.filterwarnings('ignore')

for some reason, initial efforts into the cohort analysis aren't working as well as hoped. \
so we try and dig into what proportion of these samples aren't following guidelines.

cohort = pd.read_csv('reports/test/selected-samples-no-previous-treatment-metadata.csv')
#received first line targeted therapy
cohort['FIRST_LINE_TREATMENT_TYPE'] = cohort['FIRST_LINE_TREATMENT_TYPE'].apply(lambda x: ast.literal_eval(x) if pd.notnull(x) else [])
cohort = cohort[cohort.FIRST_LINE_TREATMENT_TYPE.apply(lambda x: 'Targeted' in x)]
cohort.CANCER_TYPE.value_counts()

----

### pancreatic cancer

the only FDA approved pancreatic cancer treatment is Olaparib (BRCA1/2 mut)/Zenocutuzumab (NRG1 fusion)

cohort[cohort.CANCER_TYPE == 'Pancreatic Cancer']

- Everolimus is approved only for Astrocytoma, and variants in TSC1/TSC2/FLCN or PTEN del
- Larotrectinib is approved across all solid tumors, NTRK1/NTRK2/NTRK3 fusion
- Sunitinib is approved for KIT mutated GIST tumors

with open('reports/test/P-0035552.txt', 'r') as f:
    print(f.read())
